+++ jupyter-book/_toc.yml
@@ -63,7 +63,9 @@
         - file: tuning/parameter_tuning_manual_quiz
     - file: tuning/parameter_tuning_automated_index
       sections:
-        - file: python_scripts/parameter_tuning_automated
+        - file: python_scripts/parameter_tuning_grid_search
+        - file: python_scripts/parameter_tuning_randomized_search
+        - file: python_scripts/parameter_tuning_nested
         - file: python_scripts/parameter_tuning_ex_03
         - file: python_scripts/parameter_tuning_sol_03
         - file: tuning/parameter_tuning_automated_quiz
